BaiduWenku/requirements.txt
requests==2.22.0
pillow==7.2.0
bs4
BeautifulSoup==4.8.1
selenium==3.141.0
